+++ proxy/backend/core/src/main/java/org/apache/shardingsphere/proxy/backend/handler/distsql/ral/updatable/label/UnlabelComputeNodeExecutor.java
@@ -38,19 +38,16 @@ public final class UnlabelComputeNodeExecutor implements DistSQLUpdateExecutor<U
     
     @Override
     public void executeUpdate(final UnlabelComputeNodeStatement sqlStatement, final ContextManager contextManager) {
-        if (!contextManager.getComputeNodeInstanceContext().getModeConfiguration().isCluster()) {
-            return;
-        }
         String instanceId = sqlStatement.getInstanceId();
         Optional<ComputeNodeInstance> computeNodeInstance = contextManager.getComputeNodeInstanceContext().getClusterInstanceRegistry().find(instanceId);
         if (computeNodeInstance.isPresent()) {
             ClusterPersistServiceFacade clusterPersistServiceFacade = (ClusterPersistServiceFacade) contextManager.getPersistServiceFacade().getModeFacade();
             Collection<String> labels = new LinkedHashSet<>(computeNodeInstance.get().getLabels());
             if (sqlStatement.getLabels().isEmpty()) {
-                clusterPersistServiceFacade.getComputeNodePersistService().persistLabels(instanceId, Collections.emptyList());
+                clusterPersistServiceFacade.getComputeNodeService().persistLabels(instanceId, Collections.emptyList());
             } else {
                 labels.removeAll(sqlStatement.getLabels());
-                clusterPersistServiceFacade.getComputeNodePersistService().persistLabels(instanceId, new ArrayList<>(labels));
+                clusterPersistServiceFacade.getComputeNodeService().persistLabels(instanceId, new ArrayList<>(labels));
             }
         }
     }